feat: 테마 삭제 참조 무결성 제약조건 추가 - 테마 삭제를 시도할 때 해당 테마의 예약이 존재하면 테마를 삭제할 수 없도록 예외처리

+++ src/main/java/nextstep/reservation/repository/JdbcThemeRepository.java
@@ -2,6 +2,8 @@ package nextstep.reservation.repository;
 
 import lombok.RequiredArgsConstructor;
 import nextstep.reservation.entity.Theme;
+import nextstep.reservation.exception.RoomEscapeException;
+import nextstep.reservation.exception.RoomEscapeExceptionCode;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.RowMapper;
 import org.springframework.jdbc.support.GeneratedKeyHolder;
@@ -54,6 +56,11 @@ public class JdbcThemeRepository implements ThemeRepository {
 
     @Override
     public int deleteById(long id) {
+        String existsQuery = "select exists(select * from RESERVATION where theme_id = id)";
+        Boolean isExistInReservation = jdbcTemplate.queryForObject(existsQuery, Boolean.class, id);
+        if (isExistInReservation) {
+            throw new RoomEscapeException(RoomEscapeExceptionCode.RESERVATION_EXIST);
+        }
         String sql = "delete from THEME where id = ?";
         return jdbcTemplate.update(sql, id);
     }
